+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     1. Analytics Instance
 ### Bug Fixes
 1. Fix issue where PCA queried Compartment was not displayed as the top level compartment
+2. Resolve issues where documentation information was not be reloaded for resources.
 
 
 ## Version 0.38.0